move TMP_DIR to main project config from image
